flask-backend/requirements.txt
bcrypt==4.2.1
flask==3.1.0
flask-bcrypt==1.0.1
flask-cors==5.0.0
flask_jwt_extended==4.7.1
flask-pymongo==2.3.0
gunicorn==23.0.0
python-dotenv==1.0.1
